+++ lib/view_component/form/builder.rb
@@ -9,11 +9,31 @@ module ViewComponent
 
       class NotImplementedComponentError < Error; end
 
+      class NamespaceAlreadyAddedError < Error; end
+
+      class_attribute :lookup_namespaces, default: [ViewComponent::Form]
+
       class << self
-        attr_accessor :components_namespace
+        def inherited(base)
+          base.lookup_namespaces = lookup_namespaces.dup
+
+          super
+        end
+
+        def namespace(namespace)
+          if lookup_namespaces.include?(namespace)
+            raise NamespaceAlreadyAddedError, "The component namespace '#{namespace}' is already added"
+          end
+
+          lookup_namespaces.prepend namespace
+        end
       end
 
-      self.components_namespace = "ViewComponent::Form"
+      def initialize(*)
+        @__component_klass_cache = {}
+
+        super
+      end
 
       (field_helpers - %i[label check_box radio_button fields_for fields hidden_field file_field]).each do |selector|
         class_eval <<-RUBY_EVAL, __FILE__, __LINE__ + 1
@@ -165,20 +185,27 @@ module ViewComponent
       private
 
       def render_component(component_name, *args, &block)
-        component_klassname = "#{self.class.components_namespace}::#{component_name.to_s.camelize}Component"
-        component_klass     = component_klassname.safe_constantize
-
-        unless component_klass.is_a?(Class) && component_klass < ViewComponent::Base
-          raise NotImplementedComponentError, "Component #{component_klassname} doesn't exist" \
-                                              " or is not a ViewComponent::Base class"
-        end
-
-        component = component_klass.new(self, *args)
+        component = component_klass(component_name).new(self, *args)
         component.render_in(@template, &block)
       end
 
       def objectify_options(options)
         @default_options.merge(options.merge(object: @object))
+      end
+
+      def component_klass(component_name)
+        @__component_klass_cache[component_name] ||= begin
+          component_klass = self.class.lookup_namespaces.filter_map do |namespace|
+            "#{namespace}::#{component_name.to_s.camelize}Component".safe_constantize || false
+          end.first
+
+          unless component_klass.is_a?(Class) && component_klass < ViewComponent::Base
+            raise NotImplementedComponentError, "Component named #{component_name} doesn't exist" \
+                                                " or is not a ViewComponent::Base class"
+          end
+
+          component_klass
+        end
       end
     end
   end