+++ spec/internal/app/components/form/label_component.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Form
+  class LabelComponent < ViewComponent::Form::LabelComponent
+    def call
+      "my custom label"
+    end
+  end
+end