+++ spec/internal/app/components/inline_form/label_component.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module InlineForm
+  class LabelComponent < ViewComponent::Form::LabelComponent
+    def call
+      "my custom label for inline form"
+    end
+  end
+end