+++ spec/internal/app/forms/custom_form_builder.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class CustomFormBuilder < ViewComponent::Form::Builder
+  namespace Form
+end